train.requirements.txt
torch==1.11.0
torchaudio==0.11.0
wandb==0.12.19
scikit-learn==1.1.1
hydra-core==1.2.0
pandas==1.4.3
tqdm==4.64.0